# ElevatorControl-OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "paths": {
    "/api/elevators": {
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "CreateElevator",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "ListElevators",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/elevators/{id}/destinations": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "GetDestinations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/elevators/{id}/next": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "GetNextFloor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/elevators/{id}/move": {
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "MoveElevator",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/elevators/{id}/call": {
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "CallElevator",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/elevators/{id}/request": {
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "RequestFloor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorDestinationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ElevatorDirection": {
        "enum": [
          "Up",
          "Down"
        ],
        "type": "string"
      },
      "FloorCallRequest": {
        "type": "object",
        "properties": {
          "floor": {
            "type": "integer",
            "format": "int32"
          },
          "direction": {
            "$ref": "#/components/schemas/ElevatorDirection"
          }
        },
        "additionalProperties": false
      },
      "FloorDestinationRequest": {
        "type": "object",
        "properties": {
          "floor": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "FloorMoveRequest": {
        "type": "object",
        "properties": {
          "floor": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}